third_party/WebKit/LayoutTests/fast/dom/HTMLSelectElement/exceptions-expected.txt

This test should trigger the single exception on HTMLSelectElement, and verify that the message is reasonably helpful.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS el[1] = 0; threw exception TypeError: Failed to set an indexed property on 'HTMLSelectElement': The provided value is not of type 'HTMLOptionElement'..
PASS successfullyParsed is true

TEST COMPLETE
